fifodir: combine new and old methods
When used on systemd:
It's questionable whether the call to systemd-tmpfiles should happen. When building a
package, the directory should not be part of it. Systemd-tmpfiles will create it on
boot, and after package installation it should be evoked by the OS / packetmanager.

When using BSD or another Linux:
It probably also shouldn't be part of the package. Maybe we can drop the part about
creating the FIFO directory?